Test fixups
- Changed the course fixture to use a properly-formatted readable ID, which means it has + signs, which meant a test in the v1 views had issues with that due to urlencoding, so fixed that
- Added some fixtures to mock out the edX clone call and create a contract-ready course (e.g. has a course run)
- Updated tests in courses/views/v2 and b2b/api to use these fixtures
